+++ src/main/java/a_kind_delivery/domain/restaurant/RestaurantController.java
@@ -1,5 +1,7 @@
 package a_kind_delivery.domain.restaurant;
 
+import a_kind_delivery.domain.Exception.MyException;
+import a_kind_delivery.domain.food.FoodDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
@@ -15,6 +17,9 @@ private final RestaurantService restaurantService;
 
     @GetMapping(value="/restaurant/{id}")
     public RestaurantDTO findResByName(@PathVariable("id") Integer id){
+            if(id<0) {
+                throw new MyException("음수는 입력할 수 없습니다.");
+            }
         return restaurantService.findResByName(id);
     }
 
@@ -30,6 +35,9 @@ private final RestaurantService restaurantService;
 
     @DeleteMapping(value="/restaurant/{id}")
     public void deleteRestaurant(@PathVariable("id") Integer id){
+        if(id<0) {
+            throw new MyException("음수는 입력할 수 없습니다.");
+        }
         restaurantService.deleteRestaurant(id);
     }
 }
